fixed imports

+++ src/main/java/nl/avans/sagrada/model/toolcard/ToolCardFolieAandrukker.java
@@ -27,7 +27,7 @@ public class ToolCardFolieAandrukker extends ToolCard {
             PatternCardField removeDieField = die.getPatternCardField();
             patternCardField = patternCard.getPatternCardField(patternCardField.getxPos(), patternCardField.getyPos());
 
-            if(patternCardField.hasDie() == false && patternCard.checkSidesColor(patternCardField, die.getColor(), true) && patternCard.isNextToDie(patternCardField) == true){
+            if(patternCardField.hasDie() == false && patternCard.checkSidesColor(patternCardField, die.getColor(), true) && patternCard.checkSidesColor(patternCardField, die.getColor(), false) && patternCard.isNextToDie(patternCardField) == true){
                 removeDieField.setDie(null);
                 playerFrameFieldDao.removeDie(die, removeDieField, player);
